- many changes to support the new clearning Logger() system for   timestamped logs to stderr & a gui window. This is basically   done.
* also started work on figuring out why the iscan stopped working
  under ubuntu 7.04 (basil). This has something to do with float/double
  coercion with swig. I think this is now sorted out. the only problem
  (I know of so far) is something about the frame rate with the nvidia
  drivers -- still working on this.

*** THIS VERSION MAY BE UNSTABLE -- I SHOULD HAVE COMMITTED BEFORE I
*** STARTED MESSING AROUND WITH UBUNTU, BUT FORGOT!!



git-svn-id: svn+ssh://svn/auto/share/repos/LOCKDOWN/mlab-pype/pype2/trunk@236 97a921b0-85fb-0310-a969-e0731ffd8486

+++ src/dacq3/dacq.c
@@ -468,7 +468,7 @@ int dacq_eye_y(void)
   return(i);
 }
 
-int dacq_eye_params(float xgain, float ygain,
+int dacq_eye_params(double xgain, double ygain,
 		    int xoff, int yoff)
 {
   LOCK(semid);
@@ -491,6 +491,7 @@ int dacq_eye_read(int which)
     i = dacq_data->eye_y;
   }
   UNLOCK(semid);
+  //printf("[%d, %d]\n", which, i);
   return(i);
 }
 
@@ -596,7 +597,7 @@ void dacq_fixbreak_tau(int n)
   UNLOCK(semid);
 }
   
-int dacq_fixwin(int n, int cx, int cy, int radius, float vbias)
+int dacq_fixwin(int n, int cx, int cy, int radius, double vbias)
 {
   if (n < 0) {
     return(NFIXWIN);
